requirements.txt
# TalentSync Central Requirements
# This file contains all dependencies for the TalentSync project
# Individual services can install their specific dependencies from their pyproject.toml files

# ==========================================
# CORE FRAMEWORK & WEB
# ==========================================
fastapi
uvicorn[standard]
python-multipart
websockets

# ==========================================
# DATABASE & ORM
# ==========================================
sqlalchemy
alembic
asyncpg

# ==========================================
# DATA VALIDATION & CONFIGURATION
# ==========================================
pydantic
pydantic-settings
python-dotenv

# ==========================================
# AUTHENTICATION & SECURITY
# ==========================================
python-jose[cryptography]
passlib[bcrypt]

# ==========================================
# CACHING & MESSAGE BROKER
# ==========================================
aioredis
celery

# ==========================================
# AI & MACHINE LEARNING
# ==========================================
openai
pinecone
assemblyai
spacy

# ==========================================
# FILE PROCESSING
# ==========================================
pypdf
python-docx
aiofiles

# ==========================================
# HTTP CLIENT & API
# ==========================================
httpx

# ==========================================
# MONITORING & OBSERVABILITY
# ==========================================
prometheus-client

# ==========================================
# DEVELOPMENT & TESTING
# ==========================================
pytest
pytest-asyncio
pytest-cov
black
isort
flake8
mypy
pre-commit

# ==========================================
# NATURAL LANGUAGE PROCESSING MODELS
# ==========================================
# Note: Install spaCy English model separately:
# python -m spacy download en_core_web_sm
